requirements.txt
tensorflow
keras
numpy
pandas
matplotlib
scikit-learn
pillow
opencv-python
tk